Greatly optimize the part 1 of day 14
Instead of storing pairs of blocked points, represent the subsurface as
a HashMap: each key is an x coordinate and its value is a vector of
bools that represent the column of the subsurface. If true, it means
that the position is occupied by a rock unit or by a sand grain. This
makes the search for a blocked position much faster than before.